refactor: move mutations to lower components
