+++ src/data-workspace/data-entry-cell/entry-field-input.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import React, { useState } from 'react'
+import React from 'react'
 import { useContextSelection } from '../../context-selection/index.js'
 import { useMetadata } from '../../metadata/index.js'
 import {
@@ -7,7 +7,6 @@ import {
     getDataSetById,
 } from '../../metadata/selectors.js'
 import { OptionSet, FileResourceInput } from '../inputs/index.js'
-import { useDataValueMutation } from './use-data-value-mutation.js'
 import { VALUE_TYPES } from './value-types.js'
 
 const useDataValueParams = ({ deId, cocId }) => {
@@ -52,32 +51,6 @@ export function EntryFieldInput({ fieldname, dataElement: de, setSyncStatus }) {
     const [deId, cocId] = fieldname.split('.')
     const dataValueParams = useDataValueParams({ deId, cocId })
 
-    // todo: move into lower level components
-    // 1. Add dvParams to props in each
-    // a) here
-    // b) each file
-    // 2. Copy below into each non-file component:
-    // a) basic
-    // b) boolean
-    // c) long-text
-    // d) option-set
-    // e) true-only
-    const [lastSyncedValue, setLastSyncedValue] = useState()
-    const { mutate } = useDataValueMutation()
-    const syncData = (value) => {
-        // todo: Here's where an error state could be set: ('onError')
-        mutate(
-            // Empty values need an empty string
-            { ...dataValueParams, value: value || '' },
-            {
-                onSuccess: () => {
-                    setLastSyncedValue(value)
-                    setSyncStatus({ syncing: false, synced: true })
-                },
-            }
-        )
-    }
-
     // ? Give all these dataValueParams?
     // ? Give all these syncStatus?
     if (de.optionSetValue) {
@@ -85,11 +58,7 @@ export function EntryFieldInput({ fieldname, dataElement: de, setSyncStatus }) {
             <OptionSet
                 fieldname={fieldname}
                 optionSetId={de.optionSet.id}
-                lastSyncedValue={lastSyncedValue}
-                syncData={syncData}
-                // todo:
                 dataValueParams={dataValueParams}
-                // todo:
                 setSyncStatus={setSyncStatus}
             />
         )
@@ -108,11 +77,7 @@ export function EntryFieldInput({ fieldname, dataElement: de, setSyncStatus }) {
         return (
             <Input
                 fieldname={fieldname}
-                syncData={syncData}
-                lastSyncedValue={lastSyncedValue}
-                // todo:}
                 dataValueParams={dataValueParams}
-                // todo:
                 setSyncStatus={setSyncStatus}
             />
         )
