+++ src/data-workspace/inputs/basic-input.js
@@ -1,7 +1,9 @@
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useState } from 'react'
 import { useField } from 'react-final-form'
+import { useDataValueMutation } from '../data-entry-cell/use-data-value-mutation.js'
 import styles from './inputs.module.css'
+import { InputPropTypes } from './utils.js'
 
 export const withAdditionalProps = (Component, addlProps) => {
     return function InputWithAddlProps(props) {
@@ -11,15 +13,30 @@ export const withAdditionalProps = (Component, addlProps) => {
 
 export const BasicInput = ({
     fieldname,
-    syncData,
-    lastSyncedValue,
     inputType,
+    dataValueParams,
+    setSyncStatus,
 }) => {
     const { input, meta } = useField(fieldname, {
         // input type gets added to native `input` attributes:
         type: inputType,
         subscription: { value: true, dirty: true, valid: true },
     })
+    const [lastSyncedValue, setLastSyncedValue] = useState()
+    const { mutate } = useDataValueMutation()
+    const syncData = (value) => {
+        // todo: Here's where an error state could be set: ('onError')
+        mutate(
+            // Empty values need an empty string
+            { ...dataValueParams, value: value || '' },
+            {
+                onSuccess: () => {
+                    setLastSyncedValue(value)
+                    setSyncStatus({ syncing: false, synced: true })
+                },
+            }
+        )
+    }
 
     const handleBlur = () => {
         const { value } = input
@@ -39,12 +56,6 @@ export const BasicInput = ({
             }}
         />
     )
-}
-export const InputPropTypes = {
-    fieldname: PropTypes.string.isRequired,
-    syncData: PropTypes.func.isRequired,
-    lastSyncedValue: PropTypes.any,
-    onKeyDown: PropTypes.func,
 }
 BasicInput.propTypes = {
     ...InputPropTypes,
