+++ src/data-workspace/inputs/utils.js
@@ -12,9 +12,9 @@ export const convertCallbackSignatures = (props) => ({
 })
 
 export const InputPropTypes = {
-    name: PropTypes.string.isRequired,
-    syncData: PropTypes.func.isRequired,
+    dataValueParams: PropTypes.objectOf(PropTypes.string),
     lastSyncedValue: PropTypes.any,
+    setSyncStatus: PropTypes.func,
 }
 
 export const withAdditionalProps = (Component, addlProps) => {
